+++ CHANGELOG.md
@@ -19,7 +19,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Fixed rendering glitches in the loading screen (by @TimGoll)
 - Fixed weapon pickup through walls (by @MrXonte)
 - Fixed spectating player still being visible through thermalvision after killing that player (by @MrXonte)
-- Fixed console error when dropping ammo for weapons with no AmmoEnt. (by @MrXonte)
+- Fixed console error when dropping ammo for weapons with no AmmoEnt (by @MrXonte)
 
 ### Changed